backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
google-generativeai